Simplify plan IR: merge 12 files → 6, delete graph_builder, unify renderers, add real MOAR e2e test
- Merge plan package: nodes.py+plan.py → ir.py, schema.py+validate.py → analysis.py,
  rules/pushdown.py → rewrite.py, inline prepare.py into __init__.py, delete explain.py
- Delete graph_builder.py: inline build_operation_graph into containers.py,
  compute_operation_hashes into runner.py
- Unify plan renderers: delete format_plan (plain-text), keep format_query_plan (Rich);
  add strip_markup() for MOAR agent prompts
- Runner now lifts config once and stores plan; syntax_check/list_pipeline_operations
  walk plan nodes instead of re-parsing config
- Add real end-to-end MOAR test with 3 models (gpt-4o-mini, gpt-5-nano, gpt-5-mini)

Net: -180 lines, 6 plan files instead of 12.
