+++ docetl/display.py
@@ -9,95 +9,115 @@ from rich.markup import escape
 from docetl.agents import get_agent_tool_names
 
 if TYPE_CHECKING:
-    from docetl.containers import OpContainer
+    from docetl.plan.ir import LogicalPlan
 
 
 def format_query_plan(
-    root: OpContainer,
-    op_container_map: dict[str, OpContainer],
-    show_boundaries: bool = False,
+    plan: "LogicalPlan",
     default_model: str = "?",
 ) -> tuple[dict[str, str], str]:
-    from docetl.containers import StepBoundary
     from docetl.operations.utils.cascade_runner import (
         cascade_oracle_model,
         format_cascade_plan_lines,
     )
+    from docetl.plan.ir import JoinNode, ScanNode
 
-    step_boundaries = sorted(
-        (op for op in op_container_map.values() if isinstance(op, StepBoundary)),
-        key=lambda x: x.name,
-    )
     colors = ["cyan", "magenta", "green", "yellow", "blue", "red"]
     step_colors = {
-        b.step_name: colors[i % len(colors)] for i, b in enumerate(step_boundaries)
+        step.name: colors[i % len(colors)] for i, step in enumerate(plan.steps)
     }
 
+    printed: set[int] = set()
+
     def _fmt(
-        op: OpContainer, indent: int = 0, label: str = "", tee: bool = False
+        node, step_name: str, indent: int = 0, label: str = "", tee: bool = False
     ) -> str:
         glyph = "├" if tee else "└"
         connector = f"{'  ' * (indent - 1)}[dim]{glyph} [/dim]" if indent else ""
 
+        has_inputs = bool(node.inputs)
         s = "  " * indent
-        guide = f"{s}[dim]│[/dim] " if op.children else f"{s}  "
+        guide = f"{s}[dim]│[/dim] " if has_inputs else f"{s}  "
 
-        if isinstance(op, StepBoundary):
-            if show_boundaries:
-                color = step_colors.get(op.step_name, "white")
-                lines = [
-                    f"{connector}[{color}][bold]{op.name}[/bold][/{color}]"
-                    f"  [dim]step_boundary[/dim]{label}",
-                ]
-                lines.extend(_fmt(c, indent + 1) for c in op.children)
-                return "\n".join(lines)
-            elif op.children:
-                return _fmt(op.children[0], indent, label, tee)
-            return ""
+        if isinstance(node, ScanNode):
+            color = step_colors.get(step_name, "white")
+            return f"{connector}[{color}][bold]scan_{node.dataset_name}[/bold][/{color}]  [dim]scan[/dim]{label}"
 
-        color = step_colors.get(op.step_name, "white")
+        if id(node) in printed:
+            color = step_colors.get(step_name, "white")
+            return f"{connector}[{color}][bold]{node.name}[/bold][/{color}]  [dim](shown above)[/dim]"
+        printed.add(id(node))
+
+        color = step_colors.get(step_name, "white")
+        config = node.op_config
         lines = [
-            f"{connector}[{color}][bold]{op.name}[/bold][/{color}]"
-            f"  [cyan]{op.config['type']}[/cyan]{label}",
+            f"{connector}[{color}][bold]{node.name}[/bold][/{color}]"
+            f"  [cyan]{node.op_type}[/cyan]{label}",
         ]
 
-        if "output" in op.config and "schema" in op.config["output"]:
+        if "output" in config and "schema" in config["output"]:
             lines.append(f"{guide}[dim]output[/dim]")
-            for field, field_type in op.config["output"]["schema"].items():
+            for field, field_type in config["output"]["schema"].items():
                 lines.append(
                     f"{guide}  [bright_white]{field}[/bright_white]"
                     f" [dim]:[/dim] {escape(str(field_type))}"
                 )
 
-        agent_tools = get_agent_tool_names(op.config.get("agent"))
+        agent_tools = get_agent_tool_names(config.get("agent"))
         if agent_tools:
             lines.append(f"{guide}[dim]agent tools[/dim]")
             for tool_name in agent_tools:
-                lines.append(f"{guide}  [bright_white]{escape(tool_name)}[/bright_white]")
+                lines.append(
+                    f"{guide}  [bright_white]{escape(tool_name)}[/bright_white]"
+                )
 
-        if op.config.get("cascade"):
-            oracle_model = cascade_oracle_model(op.config, default_model)
+        if config.get("cascade"):
+            oracle_model = cascade_oracle_model(config, default_model)
             lines.extend(
                 f"{guide}{line}"
                 for line in format_cascade_plan_lines(
-                    op.config["cascade"],
-                    op_type=op.config["type"],
+                    config["cascade"],
+                    op_type=config["type"],
                     oracle_model=oracle_model,
                 )
             )
 
-        if op.children:
-            if op.is_equijoin:
-                lines.append(
-                    _fmt(op.children[0], indent + 1, " [dim](left)[/dim]", tee=True)
+        if isinstance(node, JoinNode) and len(node.inputs) == 2:
+            left_step = _step_for(node.inputs[0])
+            right_step = _step_for(node.inputs[1])
+            lines.append(
+                _fmt(
+                    node.inputs[0],
+                    left_step,
+                    indent + 1,
+                    " [dim](left)[/dim]",
+                    tee=True,
                 )
-                lines.append(_fmt(op.children[1], indent + 1, " [dim](right)[/dim]"))
-            else:
-                lines.extend(_fmt(c, indent + 1) for c in op.children)
+            )
+            lines.append(
+                _fmt(node.inputs[1], right_step, indent + 1, " [dim](right)[/dim]")
+            )
+        else:
+            for inp in node.inputs:
+                inp_step = _step_for(inp)
+                lines.append(_fmt(inp, inp_step, indent + 1))
 
         return "\n".join(lines)
 
-    return step_colors, _fmt(root)
+    step_of_cache: dict[int, str] = {}
+    for step in plan.steps:
+        for n in step.nodes:
+            step_of_cache[id(n)] = step.name
+
+    def _step_for(node) -> str:
+        return step_of_cache.get(id(node), "?")
+
+    root = plan.root
+    if root is None:
+        return step_colors, "(empty plan)"
+
+    root_step = _step_for(root)
+    return step_colors, _fmt(root, root_step)
 
 
 def format_execution_summary(
@@ -142,3 +162,10 @@ def format_execution_summary(
         + (f"Cache: [dim]{intermediate_dir}[/dim]\n" if intermediate_dir else "")
         + f"Output: [dim]{output_path}[/dim]"
     )
+
+
+def strip_markup(text: str) -> str:
+    """Strip Rich markup tags to produce plain text."""
+    from rich.text import Text
+
+    return Text.from_markup(text).plain
