--- docetl/plan/rules/__init__.py
@@ -1,3 +1,0 @@
-from docetl.plan.rules.pushdown import LimitPushdown, SelectionPushdown
-
-__all__ = ["SelectionPushdown", "LimitPushdown"]
